Fix PHPUnit tests
There was some incorrect assumptions in the PHPUnit tests about line
lengths with multibyte characters which caused test_no_groups to fail
intermittently because randomly-generated student names sometimes
contained multi-byte characters.  It appears html_to_text() is not
multi-byte sensitive so Дарья is "longer" than Donna and so can cause
wrapping to occur at a different place.

Also changed width to 0 in the task's call to html_to_text() as
presumably this is going to be wrapped when the email is sent.